Bump all the timeouts on the flasher
When we are transfering an image in the background, all the serial port
activity goes much slower if we are saturating the the link.

A higher timeout helps, we should provide configurable timeouts in the
commandline longterm.

Related-Issue: #495
(cherry picked from commit b99d39dc444fa5d3304e893c97f105ee46523728)