j: return exit code properly
fixes the bug of j tmt always returning 0

click in standalone_mode=False mode, as we use it in j, will not raise
the click.Exception.Exit but instead return the result as an integer.

this commit looks for the integer, and raisex the exception that
can be caught on the external try block.